Improve many things

+++ t/tests/crc8one.c
@@ -5,7 +5,7 @@ typedef unsigned char uint8_t;
 uint8_t crc8_one(uint8_t crc)
  {
 
-     for (int i = 0; i < 8; i++)
+     for (uint8_t i = 0; i < 8; i++)
      {
          if (crc & 0x80)
          { /* most significant bit set, shift crc register and perform XOR operation, taking not-saved 9th set bit into account */
@@ -20,6 +20,6 @@ uint8_t crc8_one(uint8_t crc)
      return crc;
  }
 
-int main(int argc, char **argv) {
+uint8_t main(uint8_t argc, char **argv) {
     return crc8_one(5);
 }
